requirements.txt
Jinja2>=3.1.0
beautifulsoup4>=4.10
edq-utils>=0.4.4
lxml>=5.3.0
lxml-stubs>=0.5.1
json5>=0.9.14
markdown-it-py>=3.0.0
mdformat>=0.7.19
mdformat_tables>=1.0.0
mdit-py-plugins>=0.4.2
requests>=2.31.0
